navigation program
this c program gets the destination point (x,y) referring to it's current location(0,0) to calculate the angular speed of left and right wheels base on the time we expect it to reach the destination. It's not a packaged function, coz I don't know the function to control the wheels.